1 exchange rate separate module
